Reduce keystone load for offline subclouds
When a large number of subclouds are offline (e.g. powered down),
the dcmanager-audit generates significant load on the keystone
services in the System Controller when it periodically (roughly
once per minute) creates a keystone token and then retrieves the
endpoints. This causes CPU usage to spike. In order to address
the issue, RegionOne client is cached in EndpointCache.

Change-Id: I17db5cd0107457f07fe14febfec23da5290153de
Story: 2007267
Task: 42082
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>