Filter out skippable alarms in precheck
Subcloud online checks are now skipped if the subcloud is
already in the 'migrated' state. The precheck also skips
upgrade alarm as well as the host lock alarm if upgrade
has started.

In addition, the commit includes the code to handle bmc_password
with None value which the previous commit
(69a74499884a8a73a3b705ce074f46202c4aa278) did not handle.

Closes-Bug: 1924774
Change-Id: Ifdfe1e44a34c2c2561c299a82d7178dce6063daf
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>