Re-initialize RegionOne keystone client properly after subcloud add
When a new subcloud is added, we expect IndexErrors while trying to
fetch the subclouds identity auth URL. At this time, the RegionOne
keystone client is also refreshed (by setting it to None). The
next subcloud that tries to access the master keystone client
cannot properly handle this and its availability is set to None for
a short duration.
This commit fixes the issue described above.

Change-Id: I3139108b119aa6b6ba20277d829c14cc61e9f88c
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1927229