Reset the force patch audit flag after it is triggered
The force patch audit flag could be set, but never cleared.
This would result in patch audits being run during every audit
loop, as well as any derived audits such as software load.

This change ensures that the force flag is cleared when a
patch audit is invoked.

The force patch audit flag is used by other audits such as the
audit_load, so it needs to be reset after those evaluations.

Change-Id: I833d157a616916a9ee37553b1f4cf3582caa8862
Closes-Bug: 1928333
Signed-off-by: albailey <Al.Bailey@windriver.com>