Updating some of the dcmanager and audit logs
On startup of any process, the entire configuration was
being logged, which was we were logging the entire
configuration to the file (almost 200 lines per process).
- Lowering that log to DEBUG.
- Added an extra 'Starting...' log entry to help indicate
when a process start/restart occurs.

The update_subcloud_endpoint_status  log was not including
enough information, as it is called several times in a row
for each endpoint.

Two audit logs that are created every 30 seconds have been reduced
from info to debug.

Closes-Bug: 1928335
Change-Id: I7ada5abf87c2f28f5826c02345f8dd3197eae665
Signed-off-by: albailey <Al.Bailey@windriver.com>