Specify the nodeset for zuul jobs
The py2.7 jobs need to specify xenial
The py3.6 jobs need to specify bionic
The focal zuul nodes only have python 3.8 installed in them

Zuul targets that invoke a generic python3 interpreter such
as pep8 is not specified.

The copyright date was updated in order to trigger
the zuul jobs, as a no-delta type of change.

Change-Id: If5352228a018f3147ae78172ae625066150c0a95
Partial-Bug: 1928978
Signed-off-by: albailey <Al.Bailey@windriver.com>