Delete existing vim strategy from subcloud during patch orch
When dcmanager creates a patch strategy, if a subcloud has an
existing vim patch strategy, it will attempt to re-use
that strategy during its patching phase, which may result in an
error.

This commit deletes the existing vim patch strategy in
a subcloud, if it exists, so it can be re-created.
If the strategy cannot be deleted, orchestration fails.

Change-Id: Id35ef26ed3ddae6d71874fc6bac11df147f72323
Closes-Bug: 1929221
Signed-off-by: albailey <Al.Bailey@windriver.com>