Adding activation retry to upgrade orchestration
When performing an activation, the keystone endpoints may not
be accessible in the subcloud due to the asyncronous way that
cert-mon can trigger a restart of keystone.

This would have occasionally resulted in the upgrade activation
failing to be initiated, and orchestration needing to be invoked
again to resume.

This 'hack' adds retries and sleeps to the initial
activation action.

Change-Id: Ic757521dec7bdc248a51a70b5463caafe7927360
Partial-Bug: 1927550
Signed-off-by: albailey <Al.Bailey@windriver.com>