Reduce strategy deletion time with large number of subclouds
When a strategy is being deleted, if the delete fails for a
particular subcloud (e.g. due to random communication failure),
the process restarts again at subcloud1 and slowly starts going
through the same subclouds again. It eventually succeeds, but it
is extremely time-intensive.

This commit fixes the issue by handling failures when attempting
to delete the strategy and allowing unreachable subclouds to be
skipped. This ensures that we don't restart the whole process or
fail the delete if a subcloud (or more than one subcloud) is
unavailable. Additionally, strategies can be deleted in parallel
which significantly reduces the total time required.

Change-Id: Ia94de29bc76f741f7ef2e572ca73e607a177c842
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1931130