Format the IP addresses in payload before adding a subcloud
The IPv6 addresses can be represented in multiple formats. As IP
addresses are stored as text in database, ansible inventory and
overrides, this commit converts the IP addresses in payload to
standard text format of IPv6 address during adding a new subcloud.

Tested with installing and bootstrapping  a new subcloud(RVMC
configured) with the correct IPv6 address values, but with
unrecommended upper case letters and '0'. The addresses are
converted to standard format in database, ansible inventory and
overrides files.

Partial-Bug: 1931459
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I6c26e749941f1ea2597f91886ad8f7da64521f0d