Make the swact unit test for upgrade orchestration run faster
The swact state in upgrade orchestration was using a hardcoded
value of 120 seconds at the end of a successful execution of
that state.

Change the code to use a constant, which can be 'mocked' to
a smaller value during unit test execution, so the swact unit
tests no longer take over two minutes to execute.

Story: 2008943
Task: 42595
Change-Id: I917fbec69d378e474baab4b4ef50503424d7590a
Signed-off-by: albailey <Al.Bailey@windriver.com>