Add mandatory/optional parameters in subcloud reinstall API
The subcloud reinstall requires bootstrap values and sysadmin password
which are not stored in the central cloud database for bootstrapping a
subcloud after reinstall. This commit adds these mandatory values to
the subcloud reinstall API, parses these values along with the
existing install values to reinstall a subcloud. In addtion, the
deploy config values are also accepted to deploy the re-installed post
bootstrap.

Tests:
Unhappy path:
1. Reinstall an online subcloud, reinstall rejected.
2. Reinstall a subcloud without mandatory bootstrap value
"system_mode", reinstall rejected.
3. Reinstall a subcloud with "management_start_address" differs from
the value stored in database, reinstall rejected.
4. Reinstall a subcloud without image in data_install, and the
software image is not uploaded in dc-vault, reinstall rejected.

Happy path:
1. Power off a managed subcloud, reinstall this subcloud with correct
bootstrap values and deploy config, the subcloud goes "installing",
"bootstrapping" and turns online and unmanaged after deployment. After
managing this subcloud, it turns in-sync status.
2. Power off a subcloud, reinstall this subcloud with only bootstrap
values offered. After the deploy status changes to "complete", issue a
dcmanager subcloud reconfigure with its deploy config values. The
subcloud will turn online after deployment.
3. Swact the active system controller, power off a subcloud, reinstall
this subcloud on the previous standby system controller. the subcloud
is reinstalled successfully and goes online after deployment.
4. Upgrade the system controllers and subcloud controllers in a DC
system, power off a subcloud after the upgrade, reinstall the subcloud
on the upgraded system controller, the reinstall is successful, and
the subcloud goes online after deployment.
5. Power off a subcloud, manually manipulate the software
version(including the value in data_install), add an image path in
data_install, reinstall this subcloud. The reinstall is successful.
Check the data in database, the software version is corrected and the
image path is changed to the image in dc-vault.

Partial-Bug: 1932034
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I6cdfaa3d476b1c2cdd3970fdfad4a5273d1b1222