py3: Fix hashlib usage
The hashlib provides some objects. Those objects have an update function
and a constructor. In Python3 both the constructor and the update
function requires an encoded bytes object instead of a str. The str is
encoded to bytes object

In Python3 the update and constructor functions of objects provided by
hashlib requires an encoded bytes.

Use the oslo_utils safe_encode method to encode the text as bytes.

Story: 2006796
Task: 42812

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I903efa218c1241e958ed93adf3ddd902bb550aa5