Eliminate unnecessary fd creation in dcorch
With 500 AIO-SX subclouds, the file descriptor (fd) usage for
the dcorch-engine processes is higher than expected. The open
file descriptors are about 13,000. The majority of the file
descriptors (almost 99%) are idle network sockets.

This update removes statically open connections that are not
used or not needed. Since both sysinv and dbsync clients are
not session-based they will not take advantage of an already
authenticated session. Cached clients are still used for both
sysinv and dbsync clients.

The test result shows the number of fds have been reduced by
more than 94%.

Story: 2008960
Task: 42868

Signed-off-by: Tao Liu <tao.liu@windriver.com>
Change-Id: Ie30910a335d5e8caa26a2f1834794146c7555e53