Reduce interaction with central cloud keystone
In order to reduce the interaction with keystone in RegionOne,
the validate token call has been removed. Additionally, when a
ConnectFailure exception is encountered, instead of reinitializing
the RegionOne keystone client, we simply retry accessing the
auth_url/endpoint. This further reduces the calls to keystone,
thus improving the performance.

Change-Id: I824d40028cb4bceaad50f940f74e3df907534424
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1938333