Eliminate unnecessary notifications to dcmanager
Dcmanager cannot keep up with RPC messages from audit workers when a
large number of subclouds are going offline. To eliminate the
unnecessary notifications from the audit workers to the dcmanager,
this commit allows the audit worker to update the dcmanager database
if the "fail_audit_count" is the only field to update.

Test:
Tests are implemented in a large DC system (with 200 AIOSX subclouds),
patching this change in the central cloud. Simulating the subclouds go
offline by deleting the static IP routes from central cloud to
subcloud via MGMT network, and then adding these routes back to
simulate the subclouds return online.

1. With this change, the 200 subclouds can be detected as offline and
online after toggling the IP routes.
2. With this change, the RPC messages are reduced when the subclouds
turn offline. With a total of four attempts, it takes 55 seconds on
average to clear the rabbitmq message queue, compared with 1m30s on
average(also four attempts) without this patch.
3. It is still time-consuming for dcmanager to process a large amount
of RPC calls when subclouds go offline. Improve the dcmanager
responsiveness will be covered in a separate story.

Story: 2008960
Task: 43075

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I01422eab0f9df3393d69fa1a6c42bce463f54aab