Fix unit test parallelization issues
Sometimes unit tests are failing in the endpoint cache code
because another unit test has already updated the cached
keystone client reference and the token validation
fails to handle that mocked object.

The fix is to ensure that the token validation is
always mocked.  The cache also needs to be cleared
on teardown, otherwise a previous test can have
differently cached values.

Story: 2008943
Task: 43305
Change-Id: I61b3d1b60a68de77b252c7128464f846315f3a60
Signed-off-by: albailey <Al.Bailey@windriver.com>