Align fernet_key_manager call parameters with SysinvClient
This corrects an error in the fernet_key_manager parameters call into
SysinvClient to align with the method declaration.

Prior to this update, there may be logs such as:
"Fail to retrieve the master fernet keys: a float is required".

TESTS PASSED:
Verify subcloud deploy and initial manage with large number of subclouds
Verify identify sync with large number of subclouds
Verify soak tests with large number of subclouds

Change-Id: I1cd50ed5c069bc578f0d1f0184bc979c8b2bcf16
Closes-Bug: 1944394
Signed-off-by: John Kung <john.kung@windriver.com>