Add features to kube-root-ca update orchestration
Enhancements for the kube rootca update strategy.

Allow passing arguments from the DC strategy to the
VIM strategy in the subclouds for kube rootca update:
 - cert-file
 - expiry-date
 - subject

The extra_args are stored in the DB for the strategy object.

The extra_args are not directly accessible to orchestration
states, so a utility method similar to what is needed for accessing
update options is provided, so that the VIM strategy creation can
use these extra parameters if they exist.

cert-file needs to be a path that is accessible to subclouds,
which means it needs to be the vault path.

When specifying the cert-file, it will copy that value to
/opt/dc-vault/certs  and reference it from there.
That vault entry will be deleted when the strategy is deleted.

Added the unit tests for the state transitions.

Fixed an issue with the audit for kube rootca always thinking
that an audit was needed.

Test Plan:
 Tested with generating a cert using expiry-date and subject
 Tested with uploading a cert
 Tested with a cert upload followed by a swact before applying

Story: 2008675
Task: 43404
Change-Id: Ia7b7d2ce3791d74341c1a0c364234a9795688e2d
Signed-off-by: albailey <Al.Bailey@windriver.com>