Minor dcmanager unit tests cleanup
- Eliminate redundant unit tests DB setup.
Some unit tests would run migration twice, once
in the parent class, and again in the child class.

- Fix DB teardown.
If DB setup failed, such as doing DB migration,
the DB was not being cleaned up because the teardown routine
was not registered when the exception was raised.
This would break all later unit tests from being able to setup.
That fix was to register the DB cleanup before the DB setup.

ctx is a variable defined in the parent unit test class, so
no need to redefine ctxt in the child classes.

Unused and old modules were being imported in test-requirements
which was breaking older versions of pip.

Story: 2008943
Task: 43574
Change-Id: I911914bcd65c3691d80ca0d8e25c3430513bbcec
Signed-off-by: albailey <Al.Bailey@windriver.com>