Retrieve RegionOne patches before executing strategy
During patch orchestration, each thread is making a call to region one
patching agent to retrieve the list of patches in the central cloud,
which becomes a bottleneck when parallelly patching a large number of
subclouds.

To reduce the number of queries to the central cloud during
a patch orchestration, this commit retrieves RegionOne patches list
upfront before spawning the threads to execute the patch strategy.

Test:
Apply a patch strategy in parallel to patch a large number of subclouds
with this change.
- Applied a patch that was applied in the system controllers before
patch orchestration.
- Applied a patch that was not applied in the system controllers
before patch orchestration.
- Removed a patch that was removed from the system controllers before
patch orchestration.
- Removed a patch that was not removed from the system controllers
before patch orchestration.

Story: 2008960
Task: 43445
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I42269486d8e61c25548c89d614e46601c3f49bbe