Avoid synching identity groups when subclouds are running N-1 load
After the system controller is upgraded to 21.12, it tries to fetch
identity group resources from the subcloud as a part of dcorch
audit but since the subclouds are running the N-1 load, it results
in "Resource Not Found" exceptions being generated in dcorch logs.

This has no functional impact but in a larger system it would cause
the dcorch logs to rotate faster when half the subclouds are
upgraded and the rest are not.

This commit fixes the issue by excluding identity groups from the
audit resource list when the subcloud runs N-1 load.

Test Plan:
1) Verify error stack trace is not present in dcorch logs
2) Verify that failure to fetch sysinv client doesn't cause the
   thread to stop
3) Verify in a large system that the additional sysinv call
   doesn't affect the system usage

Change-Id: I0bf94fcc69f12d1536b1d9f61edb8eb1005a38b5
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1949086