Fixing dc kube rootca audit to properly filter the certificate alarms
The previous audit code would always indicate the subcloud
was out-of-sync for the kube rootca update endpoint if any
certificate alarm of type expiring or expired was raised.
ie: alarm ids 500.200, 500.210

However, the audit only should indicate out-of-sync if
that alarm was raised for kube rootca certificates.
The fix was to also check the entityID of the alarms.

Test Plan:
 PASS: subcloud in-sync with no cert alarms
 PASS: subcloud in-sync with one a ssl_ca cert alarms
 PASS: subcloud out-of-sync with one kube-rootca cert alarm
 PASS: subcloud out-of-sync with kube-rootca and ssl_ca cert alarms

Change-Id: Ib04189c9cd695ab248fae7e987980a6b8b2715d1
Closes-Bug: 1949085
Signed-off-by: albailey <Al.Bailey@windriver.com>