Add validation for kube rootca during strategy create
The validation code has been duplicated from sysinv since
there is no validation library that can be imported for
the certificate subject and expiry date.

This is a usability fix, since the code acts the same, but
the strategy will not need to wait and be applied to see
the error.

Test Plan:
 Verified invalid subject is rejected during creation
 Verified invalid subject arguments rejected during creation
 Verified invalid expiry-date is rejected during creation
 Verified expired expiry-date is rejected during creation
 Verified valid values are accepted for subject and expiry date
 Verified cert-file cannot be passed with subject or expiry_date

Change-Id: I24875dbe129eb2ad4c52b2617710277472b9a89e
Closes-Bug: 1949119
Signed-off-by: albailey <Al.Bailey@windriver.com>