Handle VIM restarts and strategy queries more robustly
If the VIM is restarted during the apply of a strategy,
the VIM API would return None instead of raising an exception.

The VIM should not have been returning None, and that
has been fixed by the depends on review in stx/nfv.

Now the only time the strategy will be None is if it
is deleted by an external actor during the apply, which
is a valid scenario to fail the strategy.

These changes add an extra None check for the strategy,
and cleanup the comments related to how and why it is
handled this way.

For the failure scenario, the error now indicates:
Timeout during recovery of apply (<strategy type>) Vim strategy.
Previously it would have incorrectly indicated the strategy
did not exist.

Test Plan:
 PASS: Verified that restarting the VIM API during a subcloud
strategy apply delays the orchestration, rather than fails it.
 PASS: Verified that restarting the VIM API repeatedly during
a subcloud strategy eventually fails the orchestration.

Story: 2009665
Task: 43880
Depends-On: https://review.opendev.org/c/starlingx/nfv/+/817115
Change-Id: I39e08263be54d81e02fe3b711912374043ad56c7
Signed-off-by: albailey <Al.Bailey@windriver.com>