Fix: DC orch re-enters completed/failed state sometimes
Sometimes DC orchestration can re-enter the same state
after it has either failed or completed. This issue can
only be observed when a large number of subcloud
orchestration is performed. As the problematic code
appears to be in orch_thread.py, it affects all types
of DC orchestration.

This commit fixes the issue described above.

Test:
Successfuly completed load import (as a part of duplex
subcloud orchestration) with large number of subclouds
without re-entering the same state.

Change-Id: I57802a07009ff50d300146869efa3ceb4c9a2749
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1953519