Support persistent_size parameter in subcloud install values
This update adds the persistent_size parameter to the
install_values.yaml file, in order to be able to configure the
partition size for Platform Backup on a subcloud.

The minimum size is 30GB. If the value of the parameter is set
to less than 30GB, then the installation will not continue and
will display an error. If the persistent_size parameter is not
included in the install_values.yaml file, the Platform Backup
must be allowed a partition size of 30GB by default. Otherwise
the size specified by persistent_size will be used.

Test Plan

PASS: Verify installation of a subcloud with persistent size
      greater than 30GB in the install_values.yaml file

PASS: Verify graceful failure of the iInstallation of a
      subcloud with persistent_size less than 30GB in the
      install_values.yaml file must fail

PASS: Verify installation of a subcloud without a
      persistent_size entry in the install_values.yaml file

PASS: Verify update and reinstall with a persistent_size greater
      than the currently configured size

PASS: Verify that installation fails when updating
      persistent_size to lesser than the currently configured
      value

PASS: Verify that installation fails when persistent_size is
      set to greater than physically available size on disk

Story: 2009799
Task : 44356

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: I89716b7118fab0bcf8da67ea5652880b004f23ca