Fix prestage orchestration details for skipped subclouds
Ensure that the 'details' field reflects the reason a subcloud is
skipped during prestage orchestration.

We use a new exception, StrategySkippedException, to signal the
OrchThread that a subcloud is being skipped. This immediately
transitions the subcloud strategy to complete, while preserving
the reason for skipping in the strategy_step details.

NOTE: this update also includes a minor update
to the API parameters file, which was missed in the
original feature commit (api-ref/source/parameters.yaml)

Test Plan:

PASS:
- normal prestage orchestration
- prestage orchestration with non-AIO-SX subcloud
    - subcloud is skipped
    - details populated and reported in details column
      of 'dcmanager strategy-step list'
    - subcloud deploy_status is returned to 'complete'
- verify feature logging

Closes-Bug: 1963967
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I3408f274e3e04410872716b718130a3a27006c36