Fix prestage playbook access to subcloud denied when host key changes
Steps to Reproduce:
1 Upgrade system controller
2 Upgrade subcloud
3 prestage subcloud for reinstallation

The issue came from “Upgrade Subcloud” rather than prestage. At the
end of upgrade, it restored the old ssh configurations including
public keys on the subclouds which cause the ssh authentication
failed in the next step.

This is a generic fix for all of the playbooks (eg. reinstall,
upgrade, prestage, etc.). This UserKnownHostsFile=/dev/null ssh
option will make sure not save the host key as trusted for added
security no matter the host key changed or not. It combined with
the existing StrictHostKeyChecking=no option will disable
the public key checking for ssh connections.

Test Plan:

PASS:
- Verify installation without failures
- Verify upgrade subclouds without any access failures
- Verify prestage without subclouds access failures
- Verify no subclouds access failures according to the given
  100% reproducible steps.

Closes-Bug: 1967798
Signed-off-by: Li Zhu <li.zhu@windriver.com>
Change-Id: I100a65845875d93666f7e6b60b92f8e9add653e4