Cache RegionOne patches on orchestrated upgrade
Due to a potentially large number of parallel requests during the
initial stages of orchestrated upgrade, the RegionOne patching API may
become overloaded, causing the upgrade to fail for some subclouds.

Add a patching cache to the upgrade strategy, so that RegionOne patches
can be retrieved by workers without needing to make repeated requests to
the client API, and potentially overloading the server.

Test Plan:

PASS:
- Apply upgrade orchestration to a large number of simplex subclouds in parallel

Closes-bug: 1968321
Signed-off-by: Gabriel Silva Trevisan <gabriel.silvatrevisan@windriver.com>
Change-Id: I45201904fe0015179e31e98f419e6d82fc6806e5