Optimize sysinv health check during prestaging
When a subcloud is being prestaged we can bypass the normal health
check and assume the subcloud is online. If the subcloud is
offline, we'll either have failed to get the sysinv_client, or
prestaging will fail and the subcloud will be marked offline
in subsequent audit.

Test Plan:

PASS:
Verify normal behaviour on a large system during prestaging operations

Delete route to subcloud just before prestaging operation
- Verify that the prestaging fails and that the subcloud goes
  offline in a timely manner

Story: 2009725
Task: 45528

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I982b2e34a47962b809c3aa5bb07f95587942677e