Copy package_checksums to <software_version>_packages_list.txt
The prestaging operation on the system controller requires the list
of packages and their checksums in a file named
"<software_version>_packages_list.txt", where <software_version> is
the version of the system controller (i.e. the new version which the
subclouds are being upgraded to).

Currently, the file is stored as package_checksums. This causes
the prestaging operation to fail.

This fix ensures that the list of packages and checksums is stored
correctly as <software_version>_packages_list.txt at
/usr/local/share/pkg-list when a subcloud is installed. The file
will then be available for future prestaging operations with the
right name.

Test Plan:

PASS: Verify that prestaging of packages for upgrade is successfully
      executed.

PASS: Verify that the file is correctly stored as
      <software_version>_packages_list.txt at
      /usr/local/share/pkg-list when adding a subcloud.

PASS: Verify that package_checksums is copied to
      /usr/local/share/pkg-list as <software_version>_packages_list.txt
      when upgrading the system controller to a higher version.

PASS: Verify that package_checksums is copied to
      /usr/local/share/pkg-list as <software_version>_packages_list.txt
      when installing a controller over pxeboot

PASS: Verify that package_checksums is copied to
      /usr/local/share/pkg-list as <software_version>_packages_list.txt
      when adding a subcloud from the system controller (using dcmanager
      add)

PASS: Verify that package_checksums is copied to
      /usr/local/share/pkg-list as <software_version>_packages_list.txt
      when installing a system controller in virtualbox from
      bootimage.iso.

Closes-Bug: 1978420
Depends-On: https://review.opendev.org/c/starlingx/metal/+/845509

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: I1018d79cb91cc3af767fa8b244be180fed70b9a7